requirements.txt
numpy
Pillow
torch
torchvision
scikit-learn
tqdm
